+++ api/User/deleteUser.php
@@ -1,9 +1,15 @@
 <?php
 require_once __DIR__ . "/../../config.php";
 
-$db = Database::getInstance();
-$stmt = $db->prepare('DELETE FROM Profil WHERE id_profil = ?');
-$stmt->execute([$_SESSION['usager']]);
-unset($_SESSION['usager']);
-header('Location: /');
-exit();
+
+if (isset($_SESSION['usager'])) {
+    $db = Database::getInstance();
+    $stmt = $db->prepare('DELETE FROM Profil WHERE id_profil = ?');
+    $stmt->execute([$_SESSION['usager']]);
+    unset($_SESSION['usager']);
+    header('Location: /');
+    exit();
+} else {
+    http_response_code(401);
+    echo "Vous devez être connecté pour effectuer cette action.";
+}
